Fixed some minor issues. Effectively now has 2 modes: From packages hosting service (simplifier) by default, which can be persisted. And from persisted.
When packages are persisted they are all added to one dependency_graph.json. If non exists its created. Else they are composed.

When packages are loaded from persistence the latest local version is obtained if no specific version is defined.
Afterwards the package information and if specified its dependencies are extracted from the dependency_graph.json.